+++ packages/job-worker/src/playout/abPlayback/__tests__/abPlayback.spec.ts
@@ -1,21 +1,13 @@
-import {
-	ABResolverConfiguration,
-	ABResolverOptions,
-	IBlueprintPieceType,
-	PieceLifespan,
-	TSR,
-} from '@sofie-automation/blueprints-integration'
+import { ABResolverOptions, IBlueprintPieceType, PieceLifespan } from '@sofie-automation/blueprints-integration'
 import { EmptyPieceTimelineObjectsBlob, PieceStatusCode } from '@sofie-automation/corelib/dist/dataModel/Piece'
 import { PieceInstancePiece, ResolvedPieceInstance } from '@sofie-automation/corelib/dist/dataModel/PieceInstance'
 import { ABSessionAssignments } from '@sofie-automation/corelib/dist/dataModel/RundownPlaylist'
 import { OnGenerateTimelineObjExt } from '@sofie-automation/corelib/dist/dataModel/Timeline'
 import { literal } from '@sofie-automation/corelib/dist/lib'
 import { protectString } from '@sofie-automation/corelib/dist/protectedString'
-import { CommonContext } from '../../../blueprints/context'
-import { AssignmentResult, resolveAbAssignmentsFromRequests, SessionRequest } from '../abPlaybackResolver'
+import { AssignmentResult, resolveAbAssignmentsFromRequests } from '../abPlaybackResolver'
 import { calculateSessionTimeRanges } from '../abPlaybackSessions'
 import { AbSessionHelper } from '../abSessionHelper'
-import { applyAbPlayerObjectAssignments } from '../applyAssignments'
 
 const POOL_NAME = 'clip'
 
@@ -397,725 +389,4 @@ describe('resolveMediaPlayers', () => {
 	})
 
 	// TODO add some tests which check lookahead
-})
-
-describe('applyMediaPlayersAssignments', () => {
-	const abSessionHelper = new AbSessionHelper([], [])
-
-	const mockGetPieceSessionId: jest.MockedFunction<typeof abSessionHelper.getPieceABSessionId> = jest.fn()
-	const mockGetObjectSessionId: jest.MockedFunction<typeof abSessionHelper.getTimelineObjectAbSessionId> = jest.fn()
-
-	const context = new CommonContext({
-		name: 'test',
-		identifier: 'test',
-	})
-
-	const abConfiguration: Pick<ABResolverConfiguration, 'timelineObjectLayerChangeRules' | 'customApplyToObject'> = {}
-
-	abSessionHelper.getPieceABSessionId = mockGetPieceSessionId
-	abSessionHelper.getTimelineObjectAbSessionId = mockGetObjectSessionId
-	beforeEach(() => {
-		mockGetPieceSessionId.mockReset().mockImplementation(() => {
-			throw new Error('Method not implemented.')
-		})
-		mockGetObjectSessionId.mockReset().mockImplementation(() => {
-			throw new Error('Method not implemented.')
-		})
-	})
-
-	test('no assignments', () => {
-		const res = applyAbPlayerObjectAssignments(abSessionHelper, context, abConfiguration, [], {}, [], POOL_NAME)
-		expect(res).toEqual({})
-	})
-
-	test('only previous assignments', () => {
-		const previousAssignments: ABSessionAssignments = {
-			abc: {
-				sessionId: 'abc',
-				playerId: 5,
-				lookahead: false,
-				_rank: 1,
-			},
-			def: {
-				sessionId: 'def',
-				playerId: 3,
-				lookahead: true,
-				_rank: 2,
-			},
-		}
-
-		const res = applyAbPlayerObjectAssignments(
-			abSessionHelper,
-			context,
-			abConfiguration,
-			[],
-			previousAssignments,
-			[],
-			POOL_NAME
-		)
-		expect(res).toEqual({})
-	})
-
-	test('object with unmatched assignments', () => {
-		const previousAssignments: ABSessionAssignments = {
-			piece0_clip_def: {
-				sessionId: 'piece0_clip_def',
-				playerId: 3,
-				lookahead: false,
-				_rank: 1,
-			},
-		}
-		const pieceInstanceId = 'piece0'
-		const partInstanceId = protectString('part0')
-
-		mockGetObjectSessionId.mockImplementation((obj, name) => `${obj.pieceInstanceId}_${name}`)
-
-		const objects = [
-			literal<OnGenerateTimelineObjExt>({
-				// This should not get assigned, as it is truely unknown, and could cause all kinds of chaos
-				id: '0',
-				layer: '0',
-				enable: {
-					start: 900,
-					duration: 1000,
-				},
-				content: {
-					deviceType: TSR.DeviceType.ABSTRACT,
-				},
-				abSessions: [
-					{
-						name: 'abc',
-						pool: POOL_NAME,
-					},
-				],
-				metaData: null,
-				pieceInstanceId: pieceInstanceId,
-				partInstanceId: partInstanceId,
-			}),
-			literal<OnGenerateTimelineObjExt>({
-				// This should get assigned, as it was previously known
-				id: '1',
-				layer: '1',
-				enable: {
-					start: 3000,
-					duration: 1000,
-				},
-				content: {
-					deviceType: TSR.DeviceType.ABSTRACT,
-				},
-				abSessions: [
-					{
-						name: 'def',
-						pool: POOL_NAME,
-					},
-				],
-				metaData: null,
-				pieceInstanceId: pieceInstanceId,
-				partInstanceId: partInstanceId,
-			}),
-		]
-
-		const res = applyAbPlayerObjectAssignments(
-			abSessionHelper,
-			context,
-			abConfiguration,
-			objects,
-			previousAssignments,
-			[],
-			POOL_NAME
-		)
-		// expect(context._getNotes()).toHaveLength(0)
-		expect(res).toMatchObject({
-			piece0_clip_def: {
-				sessionId: 'piece0_clip_def',
-				playerId: 3,
-				lookahead: false,
-			},
-		})
-	})
-
-	// TODO - more tests
-})
-
-describe('resolveAbAssignmentsFromRequests', () => {
-	const TWO_SLOTS = [1, 2]
-	const THREE_SLOTS = [1, 2, 3]
-
-	function expectGotPlayer(res: AssignmentResult, id: string, playerId: number | undefined): void {
-		const req = res.requests.find((r) => r.id === id)
-		expect(req).toBeTruthy()
-		expect(req?.playerId).toEqual(playerId)
-	}
-
-	test('No pending should do nothing', () => {
-		const requests: SessionRequest[] = [
-			// Note: these should all collide
-			{
-				id: 'a',
-				start: 1000,
-				end: undefined,
-				playerId: 1,
-			},
-			{
-				id: 'b',
-				start: 1000,
-				end: undefined,
-				playerId: 1,
-			},
-			{
-				id: 'c',
-				start: 1000,
-				end: undefined,
-				playerId: 1,
-			},
-		]
-
-		const res = resolveAbAssignmentsFromRequests(resolverOptions, TWO_SLOTS, requests, 10000)
-		expect(res).toBeTruthy()
-		expect(res.failedOptional).toEqual([])
-		expect(res.failedRequired).toEqual([])
-		expect(res.requests).toEqual(requests)
-	})
-
-	test('Very basic scenario', () => {
-		const requests: SessionRequest[] = [
-			{
-				id: 'a',
-				start: 1000,
-				end: undefined,
-			},
-			{
-				id: 'b',
-				start: 2000,
-				end: undefined,
-			},
-		]
-
-		const res = resolveAbAssignmentsFromRequests(resolverOptions, TWO_SLOTS, requests, 10000)
-		expect(res).toBeTruthy()
-		expect(res.failedOptional).toEqual([])
-		expect(res.failedRequired).toEqual([])
-		expectGotPlayer(res, 'a', 1)
-		expectGotPlayer(res, 'b', 2)
-	})
-
-	test('Very basic scenario - sticky', () => {
-		const requests: SessionRequest[] = [
-			{
-				id: 'a',
-				start: 1000,
-				end: undefined,
-				playerId: 2,
-			},
-			{
-				id: 'b',
-				start: 2000,
-				end: undefined,
-			},
-		]
-
-		const res = resolveAbAssignmentsFromRequests(resolverOptions, TWO_SLOTS, requests, 10000)
-		expect(res).toBeTruthy()
-		expect(res.failedOptional).toEqual([])
-		expect(res.failedRequired).toEqual([])
-		expectGotPlayer(res, 'a', 2)
-		expectGotPlayer(res, 'b', 1)
-	})
-
-	test('Very basic collision', () => {
-		const requests: SessionRequest[] = [
-			{
-				id: 'a',
-				start: 1000,
-				end: undefined,
-			},
-			{
-				id: 'b',
-				start: 2000,
-				end: undefined,
-			},
-			{
-				id: 'c',
-				start: 3000,
-				end: undefined,
-			},
-		]
-
-		const res = resolveAbAssignmentsFromRequests(resolverOptions, TWO_SLOTS, requests, 10000)
-		expect(res).toBeTruthy()
-		expect(res.failedOptional).toEqual([])
-		expect(res.failedRequired).toEqual(['c'])
-		expectGotPlayer(res, 'a', 1)
-		expectGotPlayer(res, 'b', 2)
-		expectGotPlayer(res, 'c', undefined)
-	})
-
-	test('Very basic collision - sticky', () => {
-		const requests: SessionRequest[] = [
-			{
-				id: 'a',
-				start: 1000,
-				end: undefined,
-				playerId: 2,
-			},
-			{
-				id: 'b',
-				start: 2000,
-				end: undefined,
-			},
-			{
-				id: 'c',
-				start: 3000,
-				end: undefined,
-				playerId: 1,
-			},
-		]
-
-		const res = resolveAbAssignmentsFromRequests(resolverOptions, TWO_SLOTS, requests, 10000)
-		expect(res).toBeTruthy()
-		expect(res.failedOptional).toEqual([])
-		expect(res.failedRequired).toEqual(['b'])
-		expectGotPlayer(res, 'a', 2)
-		expectGotPlayer(res, 'b', undefined)
-		expectGotPlayer(res, 'c', 1)
-	})
-
-	test('Immediate playback', () => {
-		const requests: SessionRequest[] = [
-			{
-				id: 'a',
-				start: 1000,
-				end: 10000,
-				playerId: 2,
-			},
-			{
-				id: 'b',
-				start: 2000,
-				end: 10500,
-				playerId: 1,
-			},
-			{
-				id: 'c',
-				start: 10900,
-				end: undefined,
-			},
-			{
-				id: 'd',
-				start: 10950,
-				end: undefined,
-			},
-		]
-
-		const res = resolveAbAssignmentsFromRequests(resolverOptions, TWO_SLOTS, requests, 10000)
-		expect(res).toBeTruthy()
-		expect(res.failedOptional).toEqual([])
-		expect(res.failedRequired).toEqual([])
-		expectGotPlayer(res, 'a', 2)
-		expectGotPlayer(res, 'b', 1)
-		expectGotPlayer(res, 'c', 1)
-		expectGotPlayer(res, 'd', 2)
-	})
-
-	test('Immediate playback - gaps before', () => {
-		const requests: SessionRequest[] = [
-			{
-				id: 'a',
-				start: 1000,
-				end: 9000,
-				playerId: 2,
-			},
-			{
-				id: 'b',
-				start: 2000,
-				end: 8500,
-				playerId: 1,
-			},
-			{
-				id: 'c',
-				start: 10900,
-				end: undefined,
-			},
-			{
-				id: 'd',
-				start: 10950,
-				end: undefined,
-			},
-		]
-
-		const res = resolveAbAssignmentsFromRequests(resolverOptions, TWO_SLOTS, requests, 10000)
-		expect(res).toBeTruthy()
-		expect(res.failedOptional).toEqual([])
-		expect(res.failedRequired).toEqual([])
-		expectGotPlayer(res, 'a', 2)
-		expectGotPlayer(res, 'b', 1)
-		expectGotPlayer(res, 'c', 1)
-		expectGotPlayer(res, 'd', 2)
-	})
-
-	test('Future request', () => {
-		const requests: SessionRequest[] = [
-			{
-				id: 'a',
-				start: 1000,
-				end: 15000,
-				playerId: 3,
-			},
-			{
-				id: 'b',
-				start: 2000,
-				end: 16000,
-				playerId: 1,
-			},
-			{
-				id: 'c',
-				start: 20000,
-				end: 40000,
-			},
-			{
-				id: 'd',
-				start: 30000,
-				end: undefined,
-				playerId: 1,
-			},
-			{
-				id: 'e',
-				start: 35000,
-				end: undefined,
-			},
-		]
-
-		const res = resolveAbAssignmentsFromRequests(resolverOptions, THREE_SLOTS, requests, 10000)
-		expect(res).toBeTruthy()
-		expect(res.failedOptional).toEqual([])
-		expect(res.failedRequired).toEqual([])
-		expectGotPlayer(res, 'a', 3)
-		expectGotPlayer(res, 'b', 1)
-		expectGotPlayer(res, 'c', 2)
-		expectGotPlayer(res, 'd', 1)
-		expectGotPlayer(res, 'e', 3)
-	})
-
-	test('Run adlib bak', () => {
-		const requests: SessionRequest[] = [
-			// current part
-			{
-				id: 'a',
-				start: 1000,
-				end: undefined,
-				playerId: 2,
-			},
-			// adlib
-			{
-				id: 'b',
-				start: 10000,
-				end: undefined,
-			},
-			// lookaheads
-			{
-				id: 'c',
-				start: Number.POSITIVE_INFINITY,
-				end: undefined,
-				playerId: 1,
-				lookaheadRank: 1,
-			},
-			{
-				id: 'd',
-				start: Number.POSITIVE_INFINITY,
-				end: undefined,
-				lookaheadRank: 2,
-			},
-		]
-
-		const res = resolveAbAssignmentsFromRequests(resolverOptions, TWO_SLOTS, requests, 10000)
-		expect(res).toBeTruthy()
-		expect(res.failedOptional).toEqual([])
-		expect(res.failedRequired).toEqual([])
-		expectGotPlayer(res, 'a', 2)
-		expectGotPlayer(res, 'b', 1)
-		expectGotPlayer(res, 'c', undefined)
-		expectGotPlayer(res, 'd', undefined)
-	})
-
-	test('Run adlib bts', () => {
-		const requests: SessionRequest[] = [
-			// current part
-			{
-				id: 'a',
-				start: 1000,
-				end: 10500,
-				playerId: 2,
-			},
-			// adlib
-			{
-				id: 'b',
-				start: 10000,
-				end: undefined,
-			},
-			// lookaheads (in order of future use)
-			{
-				id: 'c',
-				start: Number.POSITIVE_INFINITY,
-				end: undefined,
-				playerId: 1,
-				lookaheadRank: 1,
-			},
-			{
-				id: 'd',
-				start: Number.POSITIVE_INFINITY,
-				end: undefined,
-				playerId: 2,
-				lookaheadRank: 2,
-			},
-		]
-
-		const res = resolveAbAssignmentsFromRequests(resolverOptions, TWO_SLOTS, requests, 10000)
-		expect(res).toBeTruthy()
-		expect(res.failedOptional).toEqual([])
-		expect(res.failedRequired).toEqual([])
-		expectGotPlayer(res, 'a', 2)
-		expectGotPlayer(res, 'b', 1)
-		expectGotPlayer(res, 'c', 2)
-		expectGotPlayer(res, 'd', undefined)
-	})
-	test('Run adlib bts x2', () => {
-		const requests: SessionRequest[] = [
-			// current part
-			{
-				id: 'a',
-				start: 1000,
-				end: 9500,
-				playerId: 2,
-			},
-			// adlib
-			{
-				id: 'b',
-				start: 10000,
-				end: undefined,
-			},
-			// adlib
-			{
-				id: 'e',
-				start: 10000,
-				end: undefined,
-			},
-			// lookaheads (in order of future use)
-			{
-				id: 'c',
-				start: Number.POSITIVE_INFINITY,
-				end: undefined,
-				playerId: 1,
-				lookaheadRank: 1,
-			},
-			{
-				id: 'd',
-				start: Number.POSITIVE_INFINITY,
-				end: undefined,
-				playerId: 2,
-				lookaheadRank: 2,
-			},
-		]
-
-		const res = resolveAbAssignmentsFromRequests(resolverOptions, TWO_SLOTS, requests, 10000)
-		expect(res).toBeTruthy()
-		expect(res.failedOptional).toEqual([])
-		expect(res.failedRequired).toEqual([])
-		expectGotPlayer(res, 'a', 2)
-		expectGotPlayer(res, 'b', 1)
-		expectGotPlayer(res, 'e', 2)
-		expectGotPlayer(res, 'c', undefined)
-		expectGotPlayer(res, 'd', undefined)
-	})
-
-	test('Autonext run bts', () => {
-		const requests: SessionRequest[] = [
-			// current part
-			{
-				id: 'a',
-				start: 1000,
-				end: 10500,
-				playerId: 2,
-			},
-			// adlib
-			{
-				id: 'b',
-				start: 10000,
-				end: 20500,
-			},
-			// next part
-			{
-				id: 'a',
-				start: 20000,
-				end: undefined,
-				playerId: 1,
-			},
-			// lookaheads (in order of future use)
-			{
-				id: 'c',
-				start: Number.POSITIVE_INFINITY,
-				end: undefined,
-				playerId: 2,
-				lookaheadRank: 1,
-			},
-			{
-				id: 'd',
-				start: Number.POSITIVE_INFINITY,
-				end: undefined,
-				lookaheadRank: 2,
-			},
-		]
-
-		const res = resolveAbAssignmentsFromRequests(resolverOptions, TWO_SLOTS, requests, 10000)
-		expect(res).toBeTruthy()
-		expect(res.failedOptional).toEqual([])
-		expect(res.failedRequired).toEqual([])
-		expectGotPlayer(res, 'a', 2)
-		expectGotPlayer(res, 'b', 1)
-		expectGotPlayer(res, 'c', 2)
-		expectGotPlayer(res, 'd', undefined)
-	})
-
-	test('Autonext run bts in 5s', () => {
-		const requests: SessionRequest[] = [
-			// current part
-			{
-				id: 'a',
-				start: 1000,
-				end: 10500,
-				playerId: 2,
-			},
-			// adlib
-			{
-				id: 'b',
-				start: 10000,
-				end: 20500,
-			},
-			// next part
-			{
-				id: 'e',
-				start: 20000,
-				end: undefined,
-				playerId: 1,
-			},
-			// lookaheads (in order of future use)
-			{
-				id: 'c',
-				start: Number.POSITIVE_INFINITY,
-				end: undefined,
-				playerId: 2,
-				lookaheadRank: 1,
-			},
-			{
-				id: 'd',
-				start: Number.POSITIVE_INFINITY,
-				end: undefined,
-				lookaheadRank: 2,
-			},
-		]
-
-		const res = resolveAbAssignmentsFromRequests(resolverOptions, TWO_SLOTS, requests, 5000)
-		expect(res).toBeTruthy()
-		expect(res.failedOptional).toEqual([])
-		expect(res.failedRequired).toEqual([])
-		expectGotPlayer(res, 'a', 2)
-		expectGotPlayer(res, 'b', 1)
-		expectGotPlayer(res, 'e', 2)
-		expectGotPlayer(res, 'c', 1)
-		expectGotPlayer(res, 'd', undefined)
-	})
-
-	test('Preserve on-air optional over a required', () => {
-		const requests: SessionRequest[] = [
-			// current part
-			{
-				id: 'a',
-				start: 1000,
-				end: undefined,
-				playerId: 2,
-			},
-			// bak
-			{
-				id: 'b',
-				start: 5000,
-				optional: true,
-				playerId: 1,
-				end: undefined,
-			},
-			// adlib
-			{
-				id: 'c',
-				start: 10000,
-				end: undefined,
-			},
-			// lookaheads
-			{
-				id: 'd',
-				start: Number.POSITIVE_INFINITY,
-				end: undefined,
-				lookaheadRank: 1,
-			},
-			{
-				id: 'e',
-				start: Number.POSITIVE_INFINITY,
-				end: undefined,
-				lookaheadRank: 2,
-			},
-		]
-
-		const res = resolveAbAssignmentsFromRequests(resolverOptions, TWO_SLOTS, requests, 10000)
-		expect(res).toBeTruthy()
-		expect(res.failedOptional).toEqual([])
-		expect(res.failedRequired).toEqual(['c'])
-		expectGotPlayer(res, 'a', 2)
-		expectGotPlayer(res, 'b', 1)
-		expectGotPlayer(res, 'c', undefined)
-		expectGotPlayer(res, 'd', undefined)
-	})
-	test('Preserve visible lookahead assignment', () => {
-		const requests: SessionRequest[] = [
-			// current clip
-			{
-				id: 'a',
-				start: 1000,
-				end: undefined,
-				playerId: 2,
-			},
-			// previous clip
-			{
-				id: 'b',
-				start: 0,
-				playerId: 1,
-				end: 5000,
-			},
-			// lookaheads
-			{
-				id: 'd',
-				start: Number.POSITIVE_INFINITY,
-				end: undefined,
-				lookaheadRank: 1,
-			},
-			{
-				id: 'e',
-				start: Number.POSITIVE_INFINITY,
-				playerId: 3, // From before
-				end: undefined,
-				lookaheadRank: 2,
-			},
-			{
-				id: 'f',
-				start: Number.POSITIVE_INFINITY,
-				end: undefined,
-				lookaheadRank: 3,
-			},
-		]
-
-		const res = resolveAbAssignmentsFromRequests(resolverOptions, THREE_SLOTS, requests, 10000)
-		expect(res).toBeTruthy()
-		expect(res.failedOptional).toEqual([])
-		expect(res.failedRequired).toEqual([])
-		expectGotPlayer(res, 'a', 2)
-		expectGotPlayer(res, 'b', 1)
-		expectGotPlayer(res, 'd', 1)
-		expectGotPlayer(res, 'e', 3)
-		expectGotPlayer(res, 'f', undefined)
-	})
 })
